+++ changelog.md
@@ -1,3 +1,13 @@
+# 2.0.0
+
+- upgrade to geode v5
+- add click action
+- fix many issues
+
+# 1.0.0
+
+- many things sorry
+
 # v0.0.1-alpha.1
 
 - Initial release